+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+- Add `CallableTree::Node::Internal#shake` to recursively execute `CallableTree::Node::Internal#reject`, including child nodes. The child nodes that are empty because their children have been rejected will also be rejected.
+- Add `CallableTree::Node::Internal#shake!` that make destructive change.
 - Add `CallableTree::Node#outline` that may be useful for writing the specs.
 
 ## [0.2.2] - 2021-10-24